requirements.txt
sentence-transformers>=2.2.2
torch>=1.12.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
requests>=2.31.0
xgboost>=2.0.0
textstat>=0.7.3
jinja2>=3.1.2
pillow>=10.0.0
